fix: share post-hero plugin with vitest and relax image caching
Extract postHeroManifest into a shared module and register it in
vitest.config so `virtual:post-heroes` resolves under Vitest; without it
every suite importing src/lib/blog failed at module resolution.

Drop `immutable` on /images/* (max-age + stale-while-revalidate) since
those filenames are unhashed and replaced in place, so in-place image
corrections now revalidate instead of being pinned for up to a year.
